This should Fix https://issues.shopware.com/issues/PT-10011 and https://issues.shopware.com/issues/PT-6785
